Server: extract /api/notify into app/routers/notify.py
Smallest extraction yet — single endpoint, single Pydantic model,
zero shared state. The handler does its own loopback check (rather
than reaching for `_ensure_loopback`) so this router has no auth-
helper dependency on server.py at all.

server.py: 8849 → 8826 lines (-23).
pytest: 465 passed, 2 skipped (unchanged).
